Set up Handler, Page, and HTML structure
I'm reusing a HTML thingy I wrote for a little Todo app test I made.
It's inspired by how Elm handles HTML. I was considering putting the
request handlers in a controllers directory, but opted not to for now.
Maybe I'll swap them over in the future. Same thing with the Pages
folder. It could just be more intuitive to structure the app around
Models, Views, and Controllers directories.